+++ spockk-compiler-plugin/src/main/kotlin/io/github/pshevche/spockk/compilation/transformer/condition/InvalidExceptionConditionExceptionFactory.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.pshevche.spockk.compilation.transformer.condition
+
+import org.jetbrains.kotlin.backend.common.CompilationException
+import org.jetbrains.kotlin.ir.IrElement
+import org.jetbrains.kotlin.ir.declarations.IrFile
+
+internal class InvalidExceptionConditionExceptionFactory(
+  private val file: IrFile,
+  private val thenBlockIr: IrElement
+) {
+
+  fun multipleExceptionConditionsException() =
+    CompilationException(
+      "A 'then' block may only have a single exception condition",
+      file,
+      thenBlockIr
+    )
+}