+++ spockk-specs/src/test/resources/samples/compilation/source/condition/NoExceptionConditionNoWrapping.kt
@@ -1,0 +1,9 @@
+class NoExceptionConditionNoWrapping : spock.lang.Specification() {
+  fun `some feature`() {
+    io.github.pshevche.spockk.lang.`when`
+    val x = 1
+
+    io.github.pshevche.spockk.lang.then
+    true
+  }
+}